# Conflicts:
#	even-business/src/main/java/com/even/service/impl/SysMenuServiceImpl.java
#	even-io/src/main/java/com/even/io/sysMenu/response/SysMenuResponse.java

+++ even-dao/src/main/java/com/even/dao/SysDeptMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SysDeptMapper {
-    long countByExample(SysDeptExample example);
+    int countByExample(SysDeptExample example);
 
     int deleteByExample(SysDeptExample example);
